refactor: 야구게임 카운트 인터페이스 파일경로 리팩토링 * 의존성 역전 원칙에 의거하여 count 인터페이스가 의존되어야 할 대상인 GameResult 로 페키지 이동.
